Update Foreach V2 serializer
