+++ modules/core/src/main/java/org/apache/synapse/config/xml/ForEachMediatorSerializer.java
@@ -22,7 +22,6 @@ package org.apache.synapse.config.xml;
 import org.apache.axiom.om.OMElement;
 import org.apache.synapse.Mediator;
 import org.apache.synapse.mediators.builtin.ForEachMediator;
-import org.apache.synapse.mediators.v2.ForEachMediatorV2;
 
 /**
  * <p>Serialize for each mediator as below : </p>
@@ -43,84 +42,39 @@ public class ForEachMediatorSerializer extends AbstractMediatorSerializer {
 
     @Override
     protected OMElement serializeSpecificMediator(Mediator m) {
-        if (m instanceof ForEachMediator) {
-            OMElement forEachElem = fac.createOMElement("foreach", synNS);
-            saveTracingState(forEachElem, m);
-
-            ForEachMediator forEachMed = (ForEachMediator) m;
-
-            if (forEachMed.getId() != null) {
-                forEachElem.addAttribute("id", forEachMed.getId(), nullNS);
-            }
-
-            if (forEachMed.getExpression() != null) {
-                SynapsePathSerializer.serializePath(forEachMed.getExpression(),
-                        forEachElem, "expression");
-            } else {
-                handleException("Missing expression of the ForEach which is required.");
-            }
-
-            if (forEachMed.getSequenceRef() != null) {
-                forEachElem.addAttribute("sequence", forEachMed.getSequenceRef(), null);
-            } else if (forEachMed.getSequence() != null) {
-                SequenceMediatorSerializer seqSerializer = new SequenceMediatorSerializer();
-                OMElement seqElement = seqSerializer.serializeAnonymousSequence(
-                        null, forEachMed.getSequence());
-                seqElement.setLocalName("sequence");
-                forEachElem.addChild(seqElement);
-            }
-
-            serializeComments(forEachElem, forEachMed.getCommentsList());
-
-            return forEachElem;
-        } else if (m instanceof ForEachMediatorV2) {
-            OMElement forEachElem = fac.createOMElement("foreach", synNS);
-            saveTracingState(forEachElem, m);
-
-            ForEachMediatorV2 forEachMediatorV2 = (ForEachMediatorV2) m;
-
-            if (forEachMediatorV2.getCollectionExpression() != null) {
-                SynapsePathSerializer.serializePath(forEachMediatorV2.getCollectionExpression(),
-                        forEachMediatorV2.getCollectionExpression().getExpression(), forEachElem, "collection");
-            } else {
-                handleException("Missing collection of the ForEach which is required.");
-            }
-            forEachElem.addAttribute(fac.createOMAttribute(
-                    "parallel-execution", nullNS, Boolean.toString(forEachMediatorV2.getParallelExecution())));
-            if (forEachMediatorV2.isContinueWithoutAggregation()) {
-                forEachElem.addAttribute(fac.createOMAttribute(
-                        ForEachMediatorFactory.ATT_CONTINUE_WITHOUT_AGGREGATION.getLocalPart(), nullNS, "true"));
-            } else {
-                forEachElem.addAttribute(fac.createOMAttribute(
-                        ForEachMediatorFactory.ATT_UPDATE_ORIGINAL.getLocalPart(), nullNS, Boolean.toString(forEachMediatorV2.isUpdateOriginal())));
-                if (!forEachMediatorV2.isUpdateOriginal()) {
-                    forEachElem.addAttribute(fac.createOMAttribute(
-                            AbstractMediatorFactory.ATT_TARGET_VARIABLE.getLocalPart(), nullNS, forEachMediatorV2.getVariableName()));
-                    forEachElem.addAttribute(fac.createOMAttribute(
-                            AbstractMediatorFactory.RESULT_TYPE_Q.getLocalPart(), nullNS, forEachMediatorV2.getContentType()));
-                    if ("XML".equalsIgnoreCase(forEachMediatorV2.getContentType())) {
-                        forEachElem.addAttribute(fac.createOMAttribute(
-                                AbstractMediatorFactory.ATT_ROOT_ELEMENT.getLocalPart(), nullNS, forEachMediatorV2.getRootElementName()));
-                    }
-                }
-            }
-            if (forEachMediatorV2.getCounterVariable() != null) {
-                forEachElem.addAttribute(fac.createOMAttribute(
-                        "counter-variable", nullNS, forEachMediatorV2.getCounterVariable()));
-            }
-            if (forEachMediatorV2.getTarget() != null) {
-                if (forEachMediatorV2.getTarget() != null && forEachMediatorV2.getTarget().getSequence() != null) {
-                    SequenceMediatorSerializer serializer = new SequenceMediatorSerializer();
-                    serializer.serializeAnonymousSequence(forEachElem, forEachMediatorV2.getTarget().getSequence());
-                }
-            } else {
-                handleException("Missing sequence element of the ForEach which is required.");
-            }
-            serializeComments(forEachElem, forEachMediatorV2.getCommentsList());
-            return forEachElem;
-        } else {
-            handleException("Unsupported mediator passed in for serialization : " + m.getType());
-            return null;
+        if (!(m instanceof ForEachMediator)) {
+            handleException("Unsupported mediator passed in for serialization : " +
+                    m.getType());
         }
+
+        OMElement forEachElem = fac.createOMElement("foreach", synNS);
+        saveTracingState(forEachElem, m);
+
+        ForEachMediator forEachMed = (ForEachMediator) m;
+
+        if (forEachMed.getId() != null) {
+            forEachElem.addAttribute("id", forEachMed.getId(), nullNS);
+        }
+
+        if (forEachMed.getExpression() != null) {
+            SynapsePathSerializer.serializePath(forEachMed.getExpression(),
+                    forEachElem, "expression");
+        } else {
+            handleException("Missing expression of the ForEach which is required.");
+        }
+
+        if (forEachMed.getSequenceRef() != null) {
+            forEachElem.addAttribute("sequence", forEachMed.getSequenceRef(), null);
+        } else if (forEachMed.getSequence() != null) {
+            SequenceMediatorSerializer seqSerializer = new SequenceMediatorSerializer();
+            OMElement seqElement = seqSerializer.serializeAnonymousSequence(
+                    null, forEachMed.getSequence());
+            seqElement.setLocalName("sequence");
+            forEachElem.addChild(seqElement);
+        }
+
+        serializeComments(forEachElem, forEachMed.getCommentsList());
+
+        return forEachElem;
     }
 }
