+++ modules/core/src/main/java/org/apache/synapse/config/xml/MediatorSerializerFinder.java
@@ -75,6 +75,7 @@ public class MediatorSerializerFinder {
             RespondMediatorSerializer.class,
             CommentMediatorSerializer.class,
             ForEachMediatorSerializer.class,
+            ForEachMediatorV2Serializer.class,
             JSONTransformMediatorSerializer.class,
             NTLMMediatorSerializer.class,
             VariableMediatorSerializer.class,
